add rust library support
This is a hackish version to compile Rust libraries with BID. The plus
with this approach is a nice integration into the build system,
including PC files generation, etc. The downside is that no cargo is
supported and that compiling to a (C) static archive is not supported
(yet).

What works:

-   manual (from Rust point of view) dependency resolution using BID
-   bindgen integration: just drop a bindgen.h and BID will invoke
    everything which is required