prog.mk: use cargo instead of Rustc
This commit adds Cargo support from within BID. This only works with a
patched version of Rustc (ATM) containing support for l4-bender as a
linker variant.
Cargo is invoked from BID which passses all relevant arguments for
l4-bender and rustc via environment variables, but a few are passed via
command-line as well. A hackish copy/install mechanism places the
resulting binary in the normal L4(Re) bin directory.
Using Cargo, external crates from crates.io can be used as on other
platforms.
The only issue is that a few information bits need to be replicated
within the Cargo.toml and the Makefile.