new pointer-based server loop registration
When a message arrives, the kernel supplies a user-defined label to
identify the sender. This label was previously used as an index into a
vector of server implementations. This requires std/alloc support and
carries the risk of performance bottle necks when the vec is rellocated.
The new approach registers a pointer  to the object itself, calling a
callback function contained within the object just mentioned. This is
currently only a proof-of-concept, yet works. The next step is to
implement proc macro derives to ease the process.