don't derive client struct by default
The client implementation is now derived as the default method
implementation of the trait methods. This has the advantage that the
user doesn't need to specify the struct in the interface definition.

The serialisation is moved into a separate module and UTCB register
access is abstracted within the ArgAccess type.