Array and string support in l4rust framework
Rudimentary support for UTCB-based arrays and strings has been added.
The current implementation only allows Vectors and Strings, but not
their unsafe counterparts as array references and str references, since
those live within the UTCB (their heap-allocated friends are copies).
This is due to a limitation in the current macro implementation + a time
issue.