[l4-rust] Fix unsafe function calls
A relimiary fix for the broken calls to unsafe functions has been put in
place. Changes to Rust and some rebasing made some functions unsafe that
weren't supposed to be unsafe and vice versa. The new code fixes this in
an acceptable, yet priliminary fashion.

Also, this fixes warnings and old left-overs from previous refactorings.